requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pandas>=2.2.0
openpyxl>=3.1.2
pypdf>=4.0.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
chromadb>=0.4.22
sentence-transformers>=2.3.1